[stf]: moved sgx exclusive code to its own module
